Fix output capture for notification message
The go-rename tool currently outputs the success message on stdout, but the
code was only capturing stderr for the notification message. Since stderr was
empty, Atom reported an error since it doesn't allow empty notification
messages.

This patch captures both stdout and stderr into a list that is then displayed
as the notification message.

Fixes issue #9